0.5.0 dev

+++ pypgx/api/pipeline.py
@@ -6,15 +6,18 @@ multiple PyPGx actions and automatically handle semantic types.
 import shutil
 import os
 
+from .. import sdk
+
 from . import utils, plot, genotype
 
 def run_ngs_pipeline(
-    gene, output, vcf=None, panel=None, tsv=None, control_statistics=None,
+    gene, output, variants=None, depth_of_coverage=None,
+    control_statistics=None, panel=None,
     force=False, do_not_plot_copy_number=False,
     do_not_plot_allele_fraction=False
 ):
     """
-    Run NGS pipeline for target gene.
+    Run WGS pipeline for a target gene.
 
     Parameters
     ----------
@@ -22,14 +25,15 @@ def run_ngs_pipeline(
         Target gene.
     output : str
         Output directory.
-    vcf : str, optional
-        VCF file.
-    panel : str, optional
-        Reference haplotype panel.
-    tsv : str, optional
-        TSV file containing read depth (zipped or unzipped).
+    variants : str, optional
+        VCF file (zipped or unzipped).
+    depth_of_coverage : str, optional
+        Archive file or object with the semantic type
+        CovFrame[DepthOfCoverage].
     control_statistics : str or pypgx.Archive, optional
         Archive file or object with the semantic type SampleTable[Statistics].
+    panel : str, optional
+        Reference haplotype panel.
     force : bool, default : False
         Overwrite output directory if it already exists.
     do_not_plot_copy_number : bool, default: False
@@ -37,6 +41,9 @@ def run_ngs_pipeline(
     do_not_plot_allele_fraction : bool, default: False
         Do not plot allele fraction profile.
     """
+    if not utils.is_target_gene(gene):
+        raise utils.NotTargetGeneError(gene)
+
     if os.path.exists(output) and force:
         shutil.rmtree(output)
 
@@ -47,10 +54,10 @@ def run_ngs_pipeline(
     alleles = None
     cnv_calls = None
 
-    if gene_table[gene_table.Gene == gene].Variants.values[0] and vcf is not None:
-        imported_variants = utils.import_variants(gene, vcf)
+    if gene_table[gene_table.Gene == gene].Variants.values[0] and variants is not None:
+        imported_variants = utils.import_variants(gene, variants)
         imported_variants.to_file(f'{output}/imported-variants.zip')
-        phased_variants = utils.estimate_phase_beagle(imported_variants, panel)
+        phased_variants = utils.estimate_phase_beagle(imported_variants, panel=panel)
         phased_variants.to_file(f'{output}/phased-variants.zip')
         consolidated_variants = utils.create_consolidated_vcf(imported_variants, phased_variants)
         consolidated_variants.to_file(f'{output}/consolidated-variants.zip')
@@ -62,10 +69,24 @@ def run_ngs_pipeline(
                 imported_variants, path=f'{output}/allele-fraction-profile'
             )
 
-    if gene_table[gene_table.Gene == gene].SV.values[0] and tsv is not None:
+    if gene_table[gene_table.Gene == gene].SV.values[0] and depth_of_coverage is not None:
+        if isinstance(depth_of_coverage, str):
+            depth_of_coverage = sdk.Archive.from_file(depth_of_coverage)
+
+        depth_of_coverage.check('CovFrame[DepthOfCoverage]')
+
         if control_statistics is None:
-            raise ValueError('CovFrame[ReadDepth] requires SampleTable[Statistcs]')
-        read_depth = utils.import_read_depth(gene, tsv)
+            raise ValueError('SV detection requires SampleTable[Statistcs]')
+
+        if isinstance(control_statistics, str):
+            control_statistics = sdk.Archive.from_file(control_statistics)
+
+        control_statistics.check('SampleTable[Statistics]')
+
+        if depth_of_coverage.metadata['Platform'] != control_statistics.metadata['Platform']:
+            raise ValueError('Different platforms detected')
+
+        read_depth = utils.import_read_depth(gene, depth_of_coverage)
         read_depth.to_file(f'{output}/read-depth.zip')
         copy_number = utils.compute_copy_number(read_depth, control_statistics)
         copy_number.to_file(f'{output}/copy-number.zip')
@@ -79,9 +100,6 @@ def run_ngs_pipeline(
             )
 
     genotypes = genotype.call_genotypes(alleles=alleles, cnv_calls=cnv_calls)
-
     genotypes.to_file(f'{output}/genotypes.zip')
-
     results = utils.combine_results(genotypes=genotypes, alleles=alleles, cnv_calls=cnv_calls)
-
     results.to_file(f'{output}/results.zip')
